FIX Refactor, ADD Marks
Add some code descriptions, refactor code parts
